+++ examples/next-i18next/pages/_app.js
@@ -1,14 +1,10 @@
 import React from 'react'
-import App from 'next/app'
 import { appWithTranslation } from '../i18n'
 
-class MyApp extends App {
-  render() {
-    const { Component, pageProps } = this.props
-    return (
-      <Component {...pageProps} />
-    )
-  }
+const App = ({ Component, pageProps }) => {
+  return <Component {...pageProps} />
 }
 
-export default appWithTranslation(MyApp)
+const AppWithTranslation = appWithTranslation(App)
+
+export default AppWithTranslation